refactor: :building_construction: Convert to TypeScript
Also remove Giphy because it wasn't being used.
