--- config.json
@@ -1,6 +1,0 @@
-{
-  "prefix": "!",
-  "channelIds": {
-    "introduce-yourself": "740409313142767686"
-  }
-}
