+++ package.json
@@ -4,7 +4,7 @@
   "description": "Open Brewery DB Discord Bot",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "ts-eager index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -13,7 +13,11 @@
     "axios": "^0.21.1",
     "discord.js": "^12.4.1",
     "dotenv": "^8.2.0",
-    "giphy-js-sdk-core": "^1.0.6"
+    "ts-eager": "^1.1.3",
+    "typescript": "^4.2.4"
   },
-  "prettier": {}
+  "prettier": {},
+  "devDependencies": {
+    "@types/node": "^14.14.37"
+  }
 }
